+++ src/nodejs/lambda-handlers/notification-consumer.js
@@ -12,8 +12,7 @@ const msg = require('message-util');
 
 const { getTemplate, getEmailTemplate } = require('./notification-consumer/templates.js');
 
-// eslint-disable-next-line
-async function sendEmailNotification({ note, email_payload }) {
+async function sendEmailNotification({ note, emailPayload }) {
   // TODO - Add additional filter for system user messages
   // logic to add DAAC content
   // eslint-disable-next-line
@@ -23,7 +22,7 @@ async function sendEmailNotification({ note, email_payload }) {
   });
 
   // Disabled until ready for uat Testing
-  // await msg.sendEmail(users, email_payload);
+  await msg.sendEmail(users, emailPayload);
 }
 
 async function processRecord(record) {