+++ terraform/s3/variables.tf
@@ -1,1 +1,3 @@
-
+variable "environment" {
+  type = string
+}