+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 <!-- Unreleased changes can be added here. -->
+- Content update - Added "Instantaneous" and help text to file_temporal_coverage
 - Content updates for data_product_status and product_temporal_coverage
 
 ## 1.0.13 - 20230-05-12