+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Unreleased changes can be added here. -->
 - Content update - Temporal and Spatial sections
+- Content update - Changed 'Temporal Coverage Notes' to 'Temporal Information Notes'
+- Content update - Changed 'Spatial Coverage Notes' to 'Spatial Information Notes'
+- Content update - help text on Spatial Resolution question on Data Publication Form
+- Content update - Added "Instantaneous" and help text to file_temporal_coverage
+- Content updates for data_product_status and product_temporal_coverage
 
 ## 1.0.13 - 20230-05-12
 - Added GES DISC community workflow